refactor(1013): extract OfflineDeviceReporter facade (SC-001, Cat B position 44)
Extract OfflineDeviceReporter (Menu 158 offline device inventory report) from
MistHelper.py to src/reports/offline_device_reporter.py under the SC-001
lazy-import facade pattern. Every live-global access (IS_TEST_MODE, InputUtils,
APICoreFetchUtils, mistapi, apisession, DataExporter, ConfigUtils) is resolved
via lazy `mh = importlib.import_module("MistHelper")` inside the methods that
need them. MistHelper.py re-exports the class alphabetically among src.reports.*
imports; menu handler at MistHelper.py:11138 continues to reach it via the
MistHelper.OfflineDeviceReporter alias.

Local gate: ruff clean, black clean, py_compile OK, --test 66/0/131 (baseline).

Cat B position 44 (SC-001).
